workers/optimize-worker/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI and NLP dependencies
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
nltk==3.8.1
spacy==3.7.2
textblob==0.18.0

# Language models and tokenizers
openai==1.3.0
anthropic==0.7.8
google-generativeai==0.3.2
cohere==4.37

# Document processing
python-docx==1.1.0
reportlab==4.0.7
PyPDF2==3.0.1
pdfkit==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1

# Text analysis and processing
regex==2023.10.3
fuzzywuzzy==0.18.0
rapidfuzz==3.6.1
textstat==0.7.3
language-tool-python==2.7.1

# Machine Learning
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2
joblib==1.3.2

# Data processing and storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
alembic==1.13.1

# Job queue and messaging
nats-py==2.7.2
celery==5.3.4
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Environment specific
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Additional ML libraries
umap-learn==0.5.5
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Vector databases for semantic search
chromadb==0.4.18
faiss-cpu==1.7.4

# Configuration and validation
pydantic-settings==2.1.0
pyyaml==6.0.1

# ATS-specific libraries
wordcloud==1.9.3
networkx==3.2.1

# Optimization-specific
optuna==3.4.0
hyperopt==0.2.7
bayesian-optimization==1.4.3
